+++ Services/SaleAnimalService.cs
@@ -1,4 +1,5 @@
-﻿using AnimalLombard.Modals;
+﻿using System.ComponentModel.DataAnnotations;
+using AnimalLombard.Modals;
 using AnimalLombard.Repository.Interfaces;
 using AnimalLombard.Services.Interfaces;
 using AnimalLombard.Repository;
@@ -28,7 +29,17 @@ public class SaleAnimalService : IAnimalService<SaleAnimal>
             .Take(pageSize)
             .ToList();
     }
-
+    public List<SaleAnimal> ShowAllAnimals(int page = 0, int pageSize = 10)
+    {
+        if (page < 0) throw new ArgumentOutOfRangeException(nameof(page), "Page must be >= 0");
+        if (pageSize <= 0) throw new ArgumentOutOfRangeException(nameof(pageSize), "PageSize must be > 0");
+        
+        var saleAnimals = _saleAnimalRepository.FindAll();
+        return saleAnimals
+            .Skip(page * pageSize)
+            .Take(pageSize)
+            .ToList();
+    }
     public SaleAnimal? BuyAnimal(string id)
     {
         var animal = _saleAnimalRepository.FindById(id);
@@ -55,8 +66,30 @@ public class SaleAnimalService : IAnimalService<SaleAnimal>
         _saleAnimalRepository.Save(animal);
     }
 
-    public void DeleteAnimal(string id)
+    public bool AddAnimal(string? name, AnimalType animalType, string? species, string? description, decimal price,
+        Dictionary<string, string>? attributes)
     {
+        if (string.IsNullOrEmpty(name) || string.IsNullOrEmpty(species) || price <= 0) return false;
+
+        try
+        {
+            var saleAnimal = SaleAnimal.Create(name, animalType, species, description ?? "", price,
+                attributes: attributes);
+            _saleAnimalRepository.Save(saleAnimal);
+            return true;
+        }
+        catch (ValidationException ex)
+        {
+            Console.WriteLine(ex);
+            return false;
+        }
+    }
+
+    public void DeleteAnimal(string? id)
+    {
+        if (string.IsNullOrEmpty(id))
+            throw new ArgumentNullException(nameof(id));
+        
         _saleAnimalRepository.Delete(id);
     }
